refactor: OAuthService를 interface 대신 abstract class로 정의해서 중복 코드 줄이기
provider에 따른 api만 다르고 login logic 자체는 동일하기 때문에 중복 코드를 줄이고자 abstract class를 정의했다.
이 abstract class를 extends하여 api만 provider에 맞게 정의해주니 중복 코드가 현저하게 줄었다.
