+++ backend/src/auth/naver/naver.service.ts
@@ -21,17 +21,14 @@ interface requestUserInfoResponse {
 }
 
 @Injectable()
-export class NaverService implements OAuthService {
-    constructor(
-        private readonly configService: ConfigService,
-        private readonly httpService: HttpService,
-        private readonly authService: AuthService
-    ) {}
+export class NaverService extends OAuthService {
+    constructor(configService: ConfigService, httpService: HttpService, authService: AuthService) {
+        super(configService, httpService, authService, 'naver');
+    }
 
     private readonly logger = new Logger(NaverService.name);
     private readonly CLIENT_ID = this.configService.get<string>('NAVER_CLIENT_ID');
     private readonly CLIENT_SECRET = this.configService.get<string>('NAVER_CLIENT_SECRET');
-    private readonly PROVIDER = 'naver';
 
     async requestToken(autherizeCode: string) {
         const { data } = await firstValueFrom(
@@ -53,18 +50,5 @@ export class NaverService implements OAuthService {
         );
 
         return data.response.id;
-    }
-
-    async login(autherizeCode: string) {
-        const { access_token: oauthAccessToken } = await this.requestToken(autherizeCode);
-        const oauthId = await this.requestUserId(oauthAccessToken);
-
-        // oauthId가 users table에 존재하는지 확인해서 로그인 or 회원가입 처리하고 userId 가져오기
-        const userId = 1;
-        const accessToken = this.authService.signToken(userId, 'access', this.PROVIDER);
-        const refreshToken = this.authService.signToken(userId, 'refresh', this.PROVIDER);
-        // users table에 refreshToken 저장
-
-        return { accessToken, refreshToken };
     }
 }
